Module website and final project
Brendan Travers
G00253985
Web Application Development Project
For this project I created a website for shopping.
It is to be used with a web server. It uses a database found in the file: web_apps.sql

Login/Register
There are pages for a user to login or register. To login just enter a user name and password. 
To register fill in all fields on the register page and submit. Once registered the users details will be saved to a table in the database on the users table.
The user table will be referred to when you login.

Shop/Cart
There are three items on a table in the shop page. If you select an item it will be kept in local storage. 
Go to the cart to see which items were selected. It shows a table of the selected items with the total price below. To remove an item from the list click on it.
There are two event listeners in the cart. One displays selected products in a list when the page is loaded and the other removes items when they are clicked on. 

Checkout